rewrite code jquery,create module, rewrite module babelify and babel
