+++ npm-task/distr/index.html
@@ -13,8 +13,9 @@
 <body>
     <main>
         <div class="input_block">
-            <label class="text" for="weatherSearch" >Type a city</label>
-            <input type="search" id="weather_Search" name="weathe" value='tallin' placeholder="Enter the city" class="input" />
+            <label class="text" for="weatherSearch">Type a city</label>
+            <input type="search" id="weather_Search" name="weathe" value='tallin' placeholder="Enter the city"
+                class="input" />
             <button id='button'>Find</button>
         </div>
         <div class="my_and_clear_button">
@@ -22,7 +23,7 @@
             <button class="clear_button">Clear</button>
         </div>
     </main>
-    <script src="bundle.js" type="text/javascript"></script>
+    <script src="bundle.js" type="module"></script>
 </body>
 
 </html>
