--- npm-task/index.html
@@ -1,28 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <link rel="stylesheet" type="text/css" href="css.css">
-    <link href="https://fonts.googleapis.com/css?family=Roboto&display=swap" rel="stylesheet">
-    <title>Search the weather</title>
-</head>
-
-<body>
-    <main>
-        <div class="input_block">
-            <label class="text" for="weatherSearch">Type a city</label>
-            <input type="search" id="weather_Search" name="weathe" placeholder="Enter the city" class="input" />
-            <button id='button'>Find</button>
-        </div>
-        <div class="my_and_clear_button">
-            <button class="my_weather">My Weather</button>
-            <button class="clear_button">Clear</button>
-        </div>
-    </main>
-    <script src="js.js" type="text/javascript"></script>
-</body>
-
-</html>
