+++ npm-task/package.json
@@ -5,13 +5,27 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "bundle":"browserify index.js -o bundle.js"
+    "bundle": "browserify src/index.js -o distr/bundle.js",
+    "bundle2": "browserify src/index.js -o distr/bundle.js -t [ babelify --presets [ @babel/preset-env @babel/preset-es2015 ]"
   },
   "author": "Irina Zaletko",
   "license": "ISC",
   "dependencies": {
-    "browserify": "^16.5.0",
-    "http-server": "^0.11.1",
     "jquery": "^3.4.1"
+  },
+  "devDependencies": {
+    "babel": "^6.23.0",
+    "babel-cli": "^6.26.0",
+    "babel-core": "^6.26.3",
+    "babel-preset-env": "^1.7.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babelify": "^8.0.0",
+    "browserify": "^16.5.0",
+    "http-server": "^0.11.1"
+  },
+  "browserify": {
+    "transform": [
+      "babelify"
+    ]
   }
 }
